requirements.txt
asgiref==3.3.1
Django==3.1.6
pytz==2021.1
sqlparse==0.4.1
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-writable-nested==0.6.2
psycopg2-binary==2.8.6
PyJWT==2.0.0